storage_service, gossip: Do not announce bootstrapping nodes with incomplete state
Currently, bootstrapping (new node or replacing existing node) has two phases:

  1) start gossiper
  2) wait for gossip to settle
  3) set STATUS, TOKENS, etc

This is problematic when replacing a node with the same ip because
right after step 1 the old nodes will mark the replacing node as alive
and can start sending traffic to the node. However, the node is not
yet marked as doing the replacing by the old nodes, and should not
receive requests. It is not yet prepared for receiving requests
until after step 3.

To fix the problem, change the gossiper to start in the shadow mode,
so that the bootstrapping node can learn about existing nodes but
existing nodes cannot learn about the bootstrapping node. We run in
this state until after step 3. This way existing nodes will see the
replacing node as alive only after it decides to comes out of shadow,
after its state is settled.