indexed_select: add global_index_prepare_data
Add new global_index_prepare_data structure which contains the data
related to global indexes, which is computed on the prepare stage of
indexed_table_select_statement. This data can be used later by the
statement, i.e. when reading index posting list (as in this change).

This change only moves a reference to the idx_token column into this new
structure.