secondary_index: fix TOKEN handling in SELECTs
Before this change, when executing SELECTs with token restrictions
combined with global index supported restriction, the token restriction
would not be applied to the index partition slice, effectively ignoring
the token restriction entirely (as if it was always true).

This change adds this case to
get_partition_slice_for_global_index_posting_list method, to slice the
partition slice according to the given token restriction.

The restriction is computed on prepare stage and stored in
global_index_prepare_data.

This commit does not change when ALLOW FILTERING clause is required.
This functionality is only supported on new indexes (with token_v2
computed column). In case of running on a old index, the behavior is
not changed (returning the same invalid data).

Fixes #7043.